キャプチャ時のフレームレート指定をサポート。フレームレート周りの整合性を取った。その他細かい修正。

+++ src/utils/capture/stream.py
@@ -96,7 +96,7 @@ class CaptureStream:
             return AISImage.from_bytes(width, height, frame_bytes)
 
     def capture_animation(
-        self, fps: int | None = None, duration_in_sec: float | None = None
+        self, fps: float | None = None, duration_in_sec: float | None = None
     ) -> list[AISImage]:
         """
         アニメ―ション（連番静止画）をキャプチャする
